Ported the error table stuff to the PC
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@5135 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/util/et/et_name.c
@@ -18,10 +18,10 @@ static const char char_set[] =
 
 static char buf[6];
 
-const char * error_table_name(num)
-    int num;
+const char * error_table_name (num)
+    long num;
 {
-    int ch;
+    long ch;
     int i;
     char *p;
